Move Dijkstra class to seperate file
